pyproject.toml
[tool.poetry]
name = "qa-bot"
version = "0.3.0"
description = ""
authors = ["CTDS UChicago <cdis@uchicago.edu>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "2.22.0"
slackclient = "2.2.1"
PyGithub = "1.44.1"
httmock = "1.4.0"
influxdb = "^5.3.1"
jira = "^3.0.1"
cryptography = "3.1.1"
ascii_graph = "^1.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
